--- src/components/header.css
@@ -1,23 +1,0 @@
-div {
-  position: relative;
-}
-
-.header-main {
-  z-index: 1;
-}
-
-.header-sub {
-  font-size: 4em;
-
-  z-index: 0;
-  position: absolute;
-  top: -1.75em;
-  left: 0;
-  color: var(--light_sub_font);
-}
-
-@media (prefers-color-scheme: dark) {
-  .header-sub {
-    color: var(--dark_sub_font);
-  }
-}
