+++ src/components/main.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import '../styles/index.css'
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
