+++ src/styles/Splash.css
